# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bookshelf - Wattpad-style book sharing platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export interface UserProfile {
  id: number;
  clerkId: string;
  username: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  bookCount?: number;
  createdAt: string;
}

export interface UpdateUserBody {
  username?: string;
  displayName?: string;
  bio?: string;
  avatarObjectPath?: string;
}

export type BookFileFormat =
  (typeof BookFileFormat)[keyof typeof BookFileFormat];

export const BookFileFormat = {
  pdf: "pdf",
  epub: "epub",
} as const;

export interface Book {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  genre?: string | null;
  authorId: string;
  authorName: string;
  /** @nullable */
  coverObjectPath?: string | null;
  fileObjectPath: string;
  fileFormat: BookFileFormat;
  isPublic: boolean;
  downloadCount: number;
  commentCount?: number;
  inLibrary?: boolean;
  createdAt: string;
  updatedAt?: string;
}

export interface BookListResponse {
  books: Book[];
  total: number;
}

export type CreateBookBodyFileFormat =
  (typeof CreateBookBodyFileFormat)[keyof typeof CreateBookBodyFileFormat];

export const CreateBookBodyFileFormat = {
  pdf: "pdf",
  epub: "epub",
} as const;

export interface CreateBookBody {
  title: string;
  description?: string;
  genre?: string;
  coverObjectPath?: string;
  fileObjectPath: string;
  fileFormat: CreateBookBodyFileFormat;
  isPublic?: boolean;
}

export interface UpdateBookBody {
  title?: string;
  description?: string;
  genre?: string;
  coverObjectPath?: string;
  isPublic?: boolean;
}

export interface DownloadResponse {
  fileUrl: string;
}

export interface LibraryEntry {
  id: number;
  bookId: number;
  userId: string;
  savedAt: string;
  book: Book;
}

export interface LibraryListResponse {
  entries: LibraryEntry[];
}

export interface Comment {
  id: number;
  bookId: number;
  authorId: string;
  authorName: string;
  /** @nullable */
  authorAvatarUrl?: string | null;
  content: string;
  createdAt: string;
}

export interface CommentListResponse {
  comments: Comment[];
}

export interface CreateCommentBody {
  content: string;
}

export interface Message {
  id: number;
  senderId: string;
  receiverId: string;
  content: string;
  read?: boolean;
  createdAt: string;
}

export interface MessageListResponse {
  messages: Message[];
  otherUser: UserProfile;
}

export interface Conversation {
  otherUser: UserProfile;
  lastMessage: Message;
  unreadCount: number;
}

export interface ConversationListResponse {
  conversations: Conversation[];
}

export interface SendMessageBody {
  content: string;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export type ListPublicBooksParams = {
  search?: string;
  genre?: string;
  limit?: number;
  offset?: number;
};

export type GetTrendingBooksParams = {
  limit?: number;
};

export type GetUserBooksParams = {
  includePrivate?: boolean;
};