# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Bookshelf - Wattpad-style book sharing platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /books:
    get:
      operationId: listPublicBooks
      tags: [books]
      summary: List public books with optional search/filter
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: genre
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of public books
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookListResponse"

    post:
      operationId: createBook
      tags: [books]
      summary: Upload a new book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookBody"
      responses:
        "201":
          description: Book created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /books/trending:
    get:
      operationId: getTrendingBooks
      tags: [books]
      summary: Get trending books (most downloaded + commented recently)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Trending books
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookListResponse"
  /books/{bookId}:
    get:
      operationId: getBook
      tags: [books]
      summary: Get a book by ID
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Book detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

    patch:
      operationId: updateBook
      tags: [books]
      summary: Update book metadata
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookBody"
      responses:
        "200":
          description: Updated book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

    delete:
      operationId: deleteBook
      tags: [books]
      summary: Delete a book
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /books/{bookId}/download:
    post:
      operationId: downloadBook
      tags: [books]
      summary: Record a download and get the file URL
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Download URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownloadResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /users/{userId}/books:
    get:
      operationId: getUserBooks
      tags: [books]
      summary: Get books uploaded by a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: includePrivate
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: User's books
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookListResponse"

  # ── Library ───────────────────────────────────────────────────────
...
  /books/{bookId}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments on a book
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentListResponse"

    post:
      operationId: createComment
      tags: [comments]
      summary: Post a comment on a book
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentBody"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /books/{bookId}/comments/{commentId}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment (owner or book author only)
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ── Messages ──────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Users ──────────────────────────────────────
...
    Book:
      type: object
      required: [id, title, authorId, authorName, fileObjectPath, fileFormat, isPublic, downloadCount, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        genre:
          type: ["string", "null"]
        authorId:
          type: string
        authorName:
          type: string
        coverObjectPath:
          type: ["string", "null"]
        fileObjectPath:
          type: string
        fileFormat:
          type: string
          enum: [pdf, epub]
        isPublic:
          type: boolean
        downloadCount:
          type: integer
        commentCount:
          type: integer
        inLibrary:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BookListResponse:
      type: object
      required: [books, total]
      properties:
        books:
          type: array
          items:
            $ref: "#/components/schemas/Book"
        total:
          type: integer
    CreateBookBody:
      type: object
      required: [title, fileObjectPath, fileFormat]
      properties:
        title:
          type: string
        description:
          type: string
        genre:
          type: string
        coverObjectPath:
          type: string
        fileObjectPath:
          type: string
        fileFormat:
          type: string
          enum: [pdf, epub]
        isPublic:
          type: boolean
          default: true
    UpdateBookBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        genre:
          type: string
        coverObjectPath:
          type: string
        isPublic:
          type: boolean
    DownloadResponse:
      type: object
      required: [fileUrl]
      properties:
        fileUrl:
          type: string

    # ── Library ────────────────────────────────────
...
    Comment:
      type: object
      required: [id, bookId, authorId, authorName, content, createdAt]
      properties:
        id:
          type: integer
        bookId:
          type: integer
        authorId:
          type: string
        authorName:
          type: string
        authorAvatarUrl:
          type: ["string", "null"]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    CommentListResponse:
      type: object
      required: [comments]
      properties:
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
    CreateCommentBody:
      type: object
      required: [content]
      properties:
        content:
          type: string

    # ── Messages ───────────────────────────────────
...